+++ internal/pkg/httpclient/httpclient.go
@@ -1,0 +1,11 @@
+package httpclient
+
+import "net/http"
+
+type HTTPClientI interface {
+	Get(url string) (resp *http.Response, err error)
+}
+
+type HTTPClient struct {
+	Client *http.Client
+}